requirements.txt
aiosqlite
aiohttp
Pillow
beautifulsoup4
ollama
discord.py==2.5.2
yt-dlp==2025.6.25
discord-ext-voice-recv==0.5.2a179
gTTS==2.5.1
faster-whisper==1.1.1
qrcode
python-barcode[images]
treys
feedparser
matplotlib
typing_extensions>=4.9
fastapi
pytest

pytest-asyncio
python-dotenv
networkx
aio_pika
textblob
httpx<0.28
playwright
uvicorn
httpx<0.28
diffusers
torch
